Fix VisionFive 2 boot and Add Star64 Board
Several issues with the current visionfive 2 support:
* JH7110 Support is only in OpenSBI Master branch, (upcoming 1.3 release).  So pin OpenSBI Revision to the commit that adds JH7110 Support otherwise reset etc does not work
* There are two variants of VisionFive2 and Star64 - 4Gb and 8Gb. StarFive's u-boot reads the eeprom on board to determine what variant you have, and modifies the loaded FDT it passes to the kernel with the correct Memory. Current U-boot scripts in meta-riscv do not do this, and thus all boards end up with the default 4Gb configuration. I've updated this for Star64, but needs testing on VF2

* Some of the addresses used in uboot for kernel/initramfs or fdt were different from upstream. Aligned the addresses with upstream StarFive.

As both Star64 and VisionFive2 are based on the same SOC, i've changed the machine configuration so we have a JH7110 SOC and VisionFive2 and Star64 inherit from this. Where appropriate in the recipies, I've changed to jh7110 instead of visionfive2

Finally, as the Star64 has different kernel/u-boot configs/dts files, i've pointed those at my repos for now, and will be upstreaming (assuming Starfive accept star64 related PRs). Once upstreamed, I'll submit a new PR here.

This also lays the groundwork for the new PineTab-V from pine64 which is due to be shipped in May.

Signed-Off-By: Justin Hammond <justin@dynam.ac>